Replace Builder with HttpClient in SessionStortage
In earlier versions of JVoiceXML, the SessionStorage returned a single
HttpClient per session; recently that was changed so that a builder is
created per session and a new HttpClient is created for each request.

The (probably unintentional) side effect of this is that no state is
retained for a session, meaning things like Cookies etc are not carried
over to each new request.