Prevent #form references from reloading a page
If a VoiceXml element (such as a <subdialog>) references a local form
(e.g. `expr='#myform'`) then the entire page will be re-fetched in order
to resolve this link. ]

However, this doesn't work when the page is dynamic. If, for example,
the VXML was generated by a POST request or if the form ID is
generated fresh on each new request, the transition into the subdialog
will fail.